requirements.txt
# Core dependencies
requests>=2.31.0
pandas>=2.0.0
numpy>=1.24.0
sodapy>=2.2.0

# Data validation and configuration
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0

# Logging
structlog>=24.0.0

# Excel processing
openpyxl>=3.1.0

# Database (Phase 2)
SQLAlchemy>=2.0.0
psycopg2-binary>=2.9.0
alembic>=1.13.0
GeoAlchemy2>=0.15.0
SQLAlchemy-Utils>=0.42.0

# Workflow orchestration (Phase 2)
apache-airflow>=2.8.0

# API (Phase 3A)
fastapi>=0.100.0
uvicorn[standard]>=0.30.0

# Frontend (Phase 3B)
streamlit>=1.30.0
plotly>=5.20.0
pydeck>=0.9.0

# Authentication & Security (Phase 3C)
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.0
python-multipart>=0.0.9

# Caching (Phase 2 & 3C)
redis>=5.0.0
hiredis>=3.0.0

# Machine Learning (Phase 3C)
scikit-learn>=1.3.0
xgboost>=2.0.0
joblib>=1.3.0

# Testing
pytest>=8.0.0
pytest-cov>=4.0.0
pytest-postgresql>=6.0.0
Faker>=20.0.0
coverage>=7.0.0
